+++ Source/Frontend/UI/UI.csproj
@@ -434,6 +434,7 @@
     <Compile Include="Components\DataGridViewNumericUpDownCell.cs" />
     <Compile Include="Components\DataGridViewNumericUpDownColumn.cs" />
     <Compile Include="Components\DataGridViewNumericUpDownEditingControl.cs" />
+    <Compile Include="Components\NumericTextBox.cs" />
     <Compile Include="UI_Extensions.cs" />
     <Compile Include="UI_VanguardImplementation.cs" />
     <EmbeddedResource Include="Components\Blast Editor\RTC_ColumnSelector_Form.resx">
